필기/백엔드/REST_API_PJT/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2.8
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
Faker==24.9.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing-extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
